truthlens/requirements.txt
fastapi
uvicorn
torch
transformers
scikit-learn
pandas
python-dotenv
shap
chromadb
sentence-transformers
